[WIP] Simplify relation registry graph
This is a rather huge change and not completed at all but it looks like
even very complex multi-through joins work (on relation level).

Here's a sum up of those changes:

  * RelationRegistry inherits from base Graph class it holds edges
    (RelationConnector) and nodes (RelationNode)

  * RelationNode has name and an alias set needed for joins

  * RelationConnector has a relationship and left and right relation
    nodes that are used for join

  * RelationConnector::Builder is used to create nodes and connectors
    (edges) - if a relationship is created via other relationship then
    the builder creates a new connector and registers its relation as
    normal node so it can be used later for other connectors

  * RelationConnector::Builder creates "via" connectors recursively so
    it can go deep into multiple-through relationships and create
    correct joined relations with correct aliasing

This is a total mess right now but I think the hardest problem is
solved.

What's left to do:

  * Finish mapping part for joined and aliased relations

  * Create a nice interface for RelationRegistry so it's easy to walk
    through the graph when needed

  * Refactor code and add unit tests

ps. I left previous implementation in data_mapper/__relation_registry/
directory for reference

Note that this is heavily based on @snusnu's work and @mbj's suggestion
about simpler graph representation.

Note about aliasing:

Aliasing is simplfied here - there's no need for building unique aliases
for all nodes...and even if there is we can add it later and use this
"basic" implementation as the foundation. We simply prefix all the
fields with the singular relation name (like user_name, song_title
etc.).