Don't inject the engine into Relation::Graph#initialize
This was previously only used to determine the
#relation_node_class and #relation_edge_class to
use within the graph. Atm, this functionality is
not needed anymore.

I've changed Relation::Graph#initialize to support
injecting those two classes instead. However, I'm
not sure if that's really needed eventually. It's
not unlikely that this will be removed or changed
in some other not yet foreseeable way.