Add test for Relation#wrap
This is mostly added as a sanity check with a reference implementation
of a mapper which can load a wrapped relation. This will be probably
removed once morpher integration is done and things settle down.

Notice that #wrap is proxied down to axiom so rom itself doesn't even
implement this.