Use block in Hash#fetch and ruin alignment
Using block is known to be a better practice (probably because we don't
allocate objects when it's not needed like this empty array).

Alignment of `=` is also against ROM's style guide which I guess should
be configured in Rubocop if it supports that (cc @chastell). FWIW I used
to be a fan of this style until I realized it's just noise and adds
friction and it also looks bad when variable name sizes differ by more
than 3 chars.