tuple_proc => row_proc
Even though "tuple" is THE word to describe data in relations we need to
differentiate "tuple" processing on the adapter layer from mapper layer
somehow.

That's why using a different word here feels appropriate. Maybe "row" is
not the best one but at least it's different. So, to summarize:

An adapter provides a dataset object that can yield "rows" which *can
be* pre-processed by a row_proc which this dataset object can provide.

A use case for it is for example CSV adapter that turns CSV::Row into a
hash (our tuple!) or YAML adapter that turns stringified hashes into
symbolized equivalents and so on. You can probably imagine more use
cases that may include some low-level data coercion. Basically anything that
feels low-level should happen on the adapter layer so that we don't need to
clutter mapper layer too much.