Add a way to define a Relation subclass
This is the first (and simple) step towards a setup where you define
your own classes and have them instantiated by ROM and stored in
registries.

The effect is exactly the same as when using the setup DSL however it
should open doors to a simpler and more obvious setup process where
you can use Ruby's object model to share behavior and knowledge across
multiple objects. For instance you should be able to define a "base"
relation class and a couple of descendant relation classes used in
different parts of your system, they would share common "core" behavior
with extra behavior specific to a given part of the system.

Please don't confuse this approach with class-orientation - there is no
class-interface that could be used at run-time. Class macros are only
used for the setup. You still end up with the env holding *objects*, not
classes.