Build adapter-specific relation subclasses
This makes it possible to inherit from ROM::Relation like that:

  class Users < ROM::Relation[:sql]
    base_name :users

    one_to_many :tasks # sql stuff is available
  end

Pros:
  - simple way of extending relations
  - no need to know the details behind adapter implementation

Cons:
  - now every adapter *must* register itself using an identifier
    because that's how we find it to extend a relation class
  - no longer possible to simply pass in a repo instance to `setup`

To improve it we could alternatively ask a repo for a relation extension
if it's available. Something like:

  if repository.relation_extension
    klass.send(:include, repository.relation_extension)
  end

Then we could still accept repo instances in setup and if adapter's
repository returns a relation extension the ROM will apply it.