Keep ThreadSafe constant in specs once loaded
This is a follow up to 6565b89 and 64f8171.

We do not depend on descendants_tracker anymore which used to require
thread_safe globally.

After removing it we only require thread_safe for memory storage.

Running (which uses memory storage)

    rspec spec/unit/rom/command_registry_spec.rb

worked only for the first spec. Other specs failed because we remove
added constants after each spec (ThreadSafe).

As a workaround exclude the ThreadSafe constant from removing.

A better solution - I think - is to just depend on thread_safe gem
again. Other gems we use already do it anyway.